Add test for product update logic. Add dropwizard specifics for run and runShadow commands to service build script.

+++ services/product/product-entities/src/main/groovy/com/myretail/domain/CurrencyPriceEntity.groovy
@@ -9,6 +9,6 @@ class CurrencyPriceEntity {
 
     CurrencyCode currency_code
 
-    BigDecimal value
+    Double value
 
 }
